+++ crates/diesel_models/src/payment_intent.rs
@@ -167,6 +167,8 @@ pub struct PaymentLinkConfigRequestForPayments {
     pub background_image: Option<PaymentLinkBackgroundImageConfig>,
     /// Custom layout for details section
     pub details_layout: Option<common_enums::PaymentLinkDetailsLayout>,
+    /// Text for payment link's handle confirm button
+    pub payment_button_text: Option<String>,
 }
 
 common_utils::impl_to_sql_from_sql_json!(PaymentLinkConfigRequestForPayments);
@@ -367,6 +369,23 @@ pub struct PaymentIntentNew {
     pub split_payments: Option<common_types::payments::SplitPaymentsRequest>,
 }
 
+#[cfg(feature = "v2")]
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum PaymentIntentUpdate {
+    /// Update the payment intent details on payment intent confirmation, before calling the connector
+    ConfirmIntent {
+        status: storage_enums::IntentStatus,
+        active_attempt_id: common_utils::id_type::GlobalAttemptId,
+        updated_by: String,
+    },
+    /// Update the payment intent details on payment intent confirmation, after calling the connector
+    ConfirmIntentPostUpdate {
+        status: storage_enums::IntentStatus,
+        amount_captured: Option<MinorUnit>,
+        updated_by: String,
+    },
+}
+
 #[cfg(feature = "v1")]
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum PaymentIntentUpdate {
@@ -512,8 +531,9 @@ pub struct PaymentIntentUpdateFields {
 #[diesel(table_name = payment_intent)]
 pub struct PaymentIntentUpdateInternal {
     pub status: Option<storage_enums::IntentStatus>,
-    pub active_attempt_id: Option<common_utils::id_type::GlobalAttemptId>,
+    pub amount_captured: Option<MinorUnit>,
     pub modified_at: PrimitiveDateTime,
+    pub active_attempt_id: Option<common_utils::id_type::GlobalAttemptId>,
     pub amount: Option<MinorUnit>,
     pub currency: Option<storage_enums::Currency>,
     pub shipping_cost: Option<MinorUnit>,