+++ crates/router/src/types/transformers.rs
@@ -1339,6 +1339,7 @@ impl ForeignFrom<storage::PaymentAttempt> for payments::PaymentAttemptResponse {
             attempt_id: payment_attempt.attempt_id,
             status: payment_attempt.status,
             amount: payment_attempt.net_amount.get_order_amount(),
+            order_tax_amount: payment_attempt.net_amount.get_order_tax_amount(),
             currency: payment_attempt.currency,
             connector: payment_attempt.connector,
             error_message: payment_attempt.error_reason,
@@ -2066,6 +2067,7 @@ impl ForeignFrom<api_models::admin::PaymentLinkConfigRequest>
             background_image: item
                 .background_image
                 .map(|background_image| background_image.foreign_into()),
+            payment_button_text: item.payment_button_text,
         }
     }
 }
@@ -2088,6 +2090,7 @@ impl ForeignFrom<diesel_models::business_profile::PaymentLinkConfigRequest>
             background_image: item
                 .background_image
                 .map(|background_image| background_image.foreign_into()),
+            payment_button_text: item.payment_button_text,
         }
     }
 }